+++ libpod/config/config_test.go
@@ -11,14 +11,14 @@ import (
 
 func TestEmptyConfig(t *testing.T) {
 	// Make sure that we can read empty configs
-	config, err := readConfigFromFile("testdata/empty.conf")
+	config, err := readConfigFromFile("testdata/empty.conf", &Config{})
 	assert.NotNil(t, config)
 	assert.Nil(t, err)
 }
 
 func TestDefaultLibpodConf(t *testing.T) {
 	// Make sure that we can read the default libpod.conf
-	config, err := readConfigFromFile("testdata/libpod.conf")
+	config, err := readConfigFromFile("testdata/libpod.conf", &Config{})
 	assert.NotNil(t, config)
 	assert.Nil(t, err)
 }
@@ -32,11 +32,8 @@ func TestMergeEmptyAndDefaultMemoryConfig(t *testing.T) {
 	defaultConfig.StateType = define.InvalidStateStore
 	defaultConfig.StorageConfig = storage.StoreOptions{}
 
-	emptyConfig, err := readConfigFromFile("testdata/empty.conf")
+	emptyConfig, err := readConfigFromFile("testdata/empty.conf", defaultConfig)
 	assert.NotNil(t, emptyConfig)
-	assert.Nil(t, err)
-
-	err = emptyConfig.mergeConfig(defaultConfig)
 	assert.Nil(t, err)
 
 	equal := reflect.DeepEqual(emptyConfig, defaultConfig)
@@ -46,17 +43,14 @@ func TestMergeEmptyAndDefaultMemoryConfig(t *testing.T) {
 func TestMergeEmptyAndLibpodConfig(t *testing.T) {
 	// Make sure that when we merge the default config into an empty one that we
 	// effectively get the default config.
-	libpodConfig, err := readConfigFromFile("testdata/libpod.conf")
+	libpodConfig, err := readConfigFromFile("testdata/libpod.conf", &Config{})
 	assert.NotNil(t, libpodConfig)
 	assert.Nil(t, err)
 	libpodConfig.StateType = define.InvalidStateStore
 	libpodConfig.StorageConfig = storage.StoreOptions{}
 
-	emptyConfig, err := readConfigFromFile("testdata/empty.conf")
+	emptyConfig, err := readConfigFromFile("testdata/empty.conf", libpodConfig)
 	assert.NotNil(t, emptyConfig)
-	assert.Nil(t, err)
-
-	err = emptyConfig.mergeConfig(libpodConfig)
 	assert.Nil(t, err)
 
 	equal := reflect.DeepEqual(emptyConfig, libpodConfig)
